datas/pocketflow-deepsearch/requirements.txt
openai>=1.0.0
duckduckgo-search>=3.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
